backend/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autogen_planner"
version = "0.1.0"
description = "AutoGen Planner with Gradio UI - Backend"
readme = "README.md"
authors = [
    {name = "AutoGen Team", email = "info@example.com"}
]
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "pydantic>=2.4.2",
    "python-dotenv>=1.0.0",
    "autogen>=0.2.0",
    "google-generativeai>=0.3.1",
    "requests>=2.31.0",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/autogen_planner"
"Bug Tracker" = "https://github.com/yourusername/autogen_planner/issues" 